Move executor helpers to separate header file
